% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib.pcss.core.R
\name{contrib.pcss.core}
\alias{contrib.pcss.core}
\title{Plot Contribution or Loadings of Traits for each Dimension/Factor from
\code{pcss.core} Output}
\usage{
contrib.pcss.core(
  x,
  ndim = NULL,
  plot.loadings = FALSE,
  use.sign = TRUE,
  sort.value = TRUE
)
}
\arguments{
\item{ndim}{The number of dimensions for which contribution or loadings of
traits are to be plotted.}

\item{plot.loadings}{If \code{TRUE}, the loadings or "right singular vectors"
are plotted instead of contributions. Default is \code{FALSE}.}

\item{use.sign}{If \code{TRUE}, contributions of variables are given the sign
of their corresponding coordinates. Default is \code{TRUE}.}

\item{sort.value}{If \code{TRUE}, the bars are sorted according to their
value.}

\item{An}{object of class \code{pcss.core}.}
}
\value{
The contributions/loadings bar plot as a \code{ggplot} object.
}
\description{
\code{screeplot.pcss.core} generates bar plots of contributions or loadings
("right singular vectors") of traits for each dimension/factor from the
output of \code{pcss.core}.
}
