% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreplot.pcss.core.R
\name{coreplot.pcss.core}
\alias{coreplot.pcss.core}
\title{Plot the cumulative variability retained by individuals/genotypes from
\code{pcss.core} Output}
\usage{
coreplot.pcss.core(x, criterion = c("size", "variance", "logistic"))
}
\arguments{
\item{x}{An object of class \code{pcss.core}.}

\item{criterion}{The core collection generation criterion. Either
\code{"size"}, \code{"variance"}, or \code{"logistic"}. See
\strong{Details}.}
}
\value{
A plot of cumulative variability retained by individuals/genotypes as
  a \code{ggplot} object. In case of \code{criterion = "logistic"}, a list
  with plots of cumulative variability retained by individuals/genotypes and
  rate of progress of cumulative contribution to variability. The size and
  variability retained by core collection are highlighted in each plot.
}
\description{
\code{coreplot.pcss.core} generates plots of cumulative variability retained
by individuals/genotypes from \code{pcss.core} Output. The size of core
collection and the corresponding cumulative variance retained are highlighted
according to the criterion used.
}
\details{
Use \code{"size"} to highlight core collection according to the threshold
\code{size} criterion or use \code{"variance"} to highlight core collection
according to the variability threshold criterion or use  \code{"logistic"} to
highlight core collection generated according to inflection point of rate of
progress of cumulative variability retained identified by logistic
regression.
}
