% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pcss.core.R
\name{print.pcss.core}
\alias{print.pcss.core}
\title{Prints summary of \code{pcss.core} object}
\usage{
\method{print}{pcss.core}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pcss.core}.}

\item{...}{Unused.}
}
\value{
The argument x, invisibly as for all \code{\link[base]{print}}
  methods.
}
\description{
\code{pcss.core} prints to console the summary of an object of class
\code{pcss.core} including the dimensionality reduction method used, the
basic details including parameters and the information on the core sets that
can be constituted.
}
\seealso{
\code{\link[rpcss]{pcss.core}}
}
