% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcss-package.R
\docType{package}
\name{rpcss-package}
\alias{rpcss}
\alias{rpcss-package}
\title{rpcss: Constitution of Core Collections by Principal Component Scoring Strategy}
\description{
Generate a Core Collection with Principal Component Scoring Strategy (PCSS) using qualitative and/or quantitative trait data according to Hamon and Noirot (1990) \url{https://www.documentation.ird.fr/hor/fdi:36506}, Noirot et al. (1996) \doi{10.2307/2527837} and Noirot et al. (2003) \url{https://www.documentation.ird.fr/hor/fdi:010031886}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aravind-j/rpcss} \url{https://cran.r-project.org/package=rpcss} \url{https://aravind-j.github.io/rpcss/} \doi{10.5281/zenodo.14889174}
  \item Report bugs at \url{https://github.com/aravind-j/rpcss/issues}
}

}
\author{
\strong{Maintainer}: J. Aravind \email{j.aravind@icar.org.in} (\href{https://orcid.org/0000-0002-4791-442X}{ORCID})

Authors:
\itemize{
  \item J. Aravind \email{j.aravind@icar.org.in} (\href{https://orcid.org/0000-0002-4791-442X}{ORCID})
  \item Anju Mahendru Singh \email{anju.singh@icar.org.in} (\href{https://orcid.org/0000-0001-6958-1630}{ORCID})
}

Other contributors:
\itemize{
  \item ICAR-NBGPR (\href{https://ror.org/00scbd467}{ROR}) (url: https://nbpgr.org.in/) [copyright holder]
}

}
\keyword{internal}
