% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.pcss.core.R
\name{biplot.pcss.core}
\alias{biplot.pcss.core}
\title{Generate Biplots from \code{pcss.core} Output}
\usage{
\method{biplot}{pcss.core}(
  x,
  ndim = 3L,
  show.traits = c("all", "none", "quantitative", "qualitative"),
  trait.scale = 1,
  point.alpha = 0.8,
  segment.alpha = 0.8
)
}
\arguments{
\item{ndim}{The number of dimensions for which biplots have to plotted.}

\item{show.traits}{Which kind of the traits to be shown in the biplot. Either
"all", "none", "quantitative" or "qualitative".}

\item{trait.scale}{A scale factor to be applied to trait coordinates plotted
in biplot.}

\item{point.alpha}{Alpha transparency value for biplot points.}

\item{segment.alpha}{Alpha transparency value for biplot segments.}

\item{An}{object of class \code{pcss.core}.}
}
\value{
A list of biplots as \code{ggplot} objects.
}
\description{
\code{biplot.pcss.core} generates biplots of scores of genotypes with or
without vectors for traits from the output of \code{pcss.core}.
}
