% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{contrib}
\alias{contrib}
\title{Generic function for \code{contrib.pcss.core}}
\usage{
contrib(x, ...)
}
\arguments{
\item{x}{An object of class \code{pcss.core}.}

\item{...}{Unused.}
}
\description{
Generic function for \code{contrib.pcss.core}
}
\keyword{internal}
