% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coreplot}
\alias{coreplot}
\title{Generic function for \code{coreplot.pcss.core}}
\usage{
coreplot(x, ...)
}
\arguments{
\item{x}{An object of class \code{pcss.core}.}

\item{...}{Unused.}
}
\description{
Generic function for \code{coreplot.pcss.core}
}
\keyword{internal}
