% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcss.core.R
\name{pcss.core}
\alias{pcss.core}
\title{Principal Component Scoring to Generate Core collections}
\usage{
pcss.core(
  data,
  names,
  quantitative,
  qualitative,
  eigen.threshold = NULL,
  size = 0.2,
  var.threshold = 0.75
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string.}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{eigen.threshold}{The lower limit of the eigen value of factors to be
included in the estimation. The default value is the average of all the
eigen values.}

\item{size}{The desired core set size proportion.}

\item{var.threshold}{The desired proportion of total variabilty to be}
}
\value{
A list of class \code{pcss.core} with the following components.
  \item{details}{The details of the core set generaton process.}
  \item{raw.out}{The original output of \code{\link[FactoMineR]{PCA}},
  \code{\link[FactoMineR]{CA}} and \code{\link[FactoMineR]{FAMD}} functions
  of \code{\link[FactoMineR]{FactoMineR}}} \item{eigen}{A data frame with
  eigen values and their partial and cumulative contribution to percentage of
  variance.} \item{eigen.threshold}{The threshold eigen value used.}
  \item{rotation}{A matrix of rotation values or loadings.} \item{scores}{A
  matrix of scores from PCA, CA or FAMD.} \item{variability.ret}{A data frame
  of genotypes ordered by variability retained.} \item{cores.info}{A data frame of core set
  size and percentage variablity retained according to the method used.}
  \item{core.plots}{A list of plots of cumulative variability retained by
  genotypes as \code{ggplot} objects. The core set size and corresponding
  percentage variability retained are highlighted according to the method
  used.}
}
\description{
Generate a Core Collection with Principal Component Scoring Strategy (PCSS)
\insertCite{noirot_principal_1996,noirot_method_2003}{rpcss} using
qualitative and/or quantitative trait data.
}
\details{
For qualitative traits, dimensionality reduction is performed by Principal
Component Analysis (PCA). Similarly for quantitative traits, dimensionality
reduction is done via Multiple Correspondence Analysis (MCA). For cases where
both quantitative and qualitative traits are available, dimensionality
reduction is achieved by using the Factor Analysis for Mixed Data (FAMD)
method \insertCite{husson_Exploratory_2017}.
}
\seealso{
\code{\link[FactoMineR]{PCA}}, \code{\link[FactoMineR]{CA}} and
  \code{\link[FactoMineR]{FAMD}}
}
