% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot.pcss.core.R
\name{screeplot.pcss.core}
\alias{screeplot.pcss.core}
\title{Plot Eigen values as a Scree Plot from \code{pcss.core} Output}
\usage{
\method{screeplot}{pcss.core}(x, ndim = NULL, show.values = TRUE)
}
\arguments{
\item{ndim}{The number of eigen values to be plotted in the scree plot.}

\item{show.values}{If \code{TRUE}, the eigen values are shown in the plot as
annotation labels. Default is \code{TRUE}.}

\item{An}{object of class \code{pcss.core}.}
}
\value{
The scree plot as a \code{ggplot} object.
}
\description{
\code{screeplot.pcss.core} generates a scree plot of eigen values from the
output of \code{pcss.core}.
}
