% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.pcss.core.R
\name{subset.pcss.core}
\alias{subset.pcss.core}
\title{Fetch the names of individuals/genotypes in the core set generated from
\code{pcss.core} Output}
\usage{
\method{subset}{pcss.core}(x, criterion = c("size", "variance", "logistic"))
}
\arguments{
\item{x}{An object of class \code{pcss.core}.}

\item{criterion}{The core collection generation criterion. Either
\code{"size"}, \code{"variance"}, or \code{"logistic"}. See
\strong{Details}.}
}
\value{
The names of individuals/genotypes in the core collection as a
  character vector.
}
\description{
\code{subset.pcss.core} returns names of individuals/genotypes in the core
collection from \code{pcss.core} Output.
}
\details{
Use \code{"size"} to return names of individuals/genotypes in the core
collection according to the threshold \code{size} criterion or use
\code{"variance"} to return names according to the variability threshold
criterion or use  \code{"logistic"} to return names according to inflection
point of rate of progress of cumulative variability retained identified by
logistic regression.
}
